Add storage-capacity option and use it instead of static radius
Static radius will cause the database to grow as the total
content on the network grows. This was fine in the beginning
to ensure enough nodes had the content but becomes impossible to
maintain now that the network is growing substancially.
Instead we now set it to a fixed max storage capacity.